Migrate to scenetest 0.12 consolidated package layout
Pure renames, no API or behavior changes:
- Replace @scenetest/checks-react dep with @scenetest/checks (^0.12.0)
- Fold @scenetest/checks-panel into @scenetest/vite-plugin (^0.12.0)
- Leave @scenetest/scenes at its current range (unchanged this release)
- Rewrite '@scenetest/checks-react' imports to '@scenetest/checks/react'
- Update declare-module block and st:link/st:unlink scripts
