+++ package.json
@@ -17,8 +17,8 @@
 		"scene": "scenetest --no-panel",
 		"scene:ui": "scenetest",
 		"scene:ci": "scenetest --no-panel --format both --report scenetest/.reports",
-		"st:link": "pnpm link ../scenetest-js/packages/scenes && pnpm link ../scenetest-js/packages/vite-plugin && pnpm link ../scenetest-js/packages/checks-panel && pnpm link ../scenetest-js/packages/checks-react",
-		"st:unlink": "pnpm unlink @scenetest/scenes @scenetest/vite-plugin @scenetest/checks-panel @scenetest/checks-react && pnpm install",
+		"st:link": "pnpm link ../scenetest-js/packages/scenes && pnpm link ../scenetest-js/packages/vite-plugin && pnpm link ../scenetest-js/packages/checks",
+		"st:unlink": "pnpm unlink @scenetest/scenes @scenetest/vite-plugin @scenetest/checks && pnpm install",
 		"test:unit": "vitest run",
 		"migrate": "supabase db diff --use-pg-delta -f new_migration",
 		"types": "supabase gen types typescript --local > ./src/types/supabase.ts",
@@ -66,10 +66,9 @@
 		"@playwright/test": "^1.58.2",
 		"@prettier/plugin-oxc": "^0.1.3",
 		"@rolldown/plugin-babel": "^0.2.3",
-		"@scenetest/checks-panel": "^0.10.0",
-		"@scenetest/checks-react": "^0.10.0",
+		"@scenetest/checks": "^0.12.0",
 		"@scenetest/scenes": "^0.10.0",
-		"@scenetest/vite-plugin": "^0.10.0",
+		"@scenetest/vite-plugin": "^0.12.0",
 		"@tailwindcss/container-queries": "^0.1.1",
 		"@tailwindcss/line-clamp": "^0.4.4",
 		"@tailwindcss/typography": "^0.5.19",
