+++ src/features/requests/collections.ts
@@ -21,7 +21,7 @@ import {
 } from './schemas'
 import { queryClient } from '@/lib/query-client'
 import supabase from '@/lib/supabase-client'
-import { should } from '@scenetest/checks-react'
+import { should } from '@scenetest/checks/react'
 
 export const phraseRequestsCollection = createCollection(
 	queryCollectionOptions({
