+++ src/types/scenetest.d.ts
@@ -2,9 +2,9 @@ import type { SupabaseClient } from '@supabase/supabase-js'
 
 // Augment scenetest's ServerContext so `server.supabase` is typed inside
 // serverCheck() callbacks. The matching runtime value is wired up in
-// scenetest/config.ts under `server.supabase`. pnpm hoists checks-react's
-// nested @scenetest/checks separately from the top-level one, so augment
-// both module specifiers.
+// scenetest/config.ts under `server.supabase`. pnpm hoists the checks/react
+// subpath's nested @scenetest/checks separately from the top-level one, so
+// augment both module specifiers.
 
 declare module '@scenetest/checks' {
 	interface ServerContext {
@@ -12,10 +12,8 @@ declare module '@scenetest/checks' {
 	}
 }
 
-declare module '@scenetest/checks-react' {
+declare module '@scenetest/checks/react' {
 	interface ServerContext {
 		supabase: SupabaseClient
 	}
 }
-
-
